+++ src/lib/draw3D/constants.ts
@@ -1,0 +1,3 @@
+import { AU } from '../constants.ts';
+
+export const SCALE_FACTOR = AU; // 1e9;
